+++ CHANGES.rst
@@ -10,6 +10,9 @@ Changelog
 - Added basic tests to package.
   [loechel]
 
+- Changed parser.py to use functions from zc.buildout o get versions and versionannotations section names.
+  [loechel]
+
 1.5.1 (2017-01-23)
 ------------------